+++ samples/java/hibernate/src/main/java/com/google/cloud/postgres/models/Concerts.java
@@ -23,7 +23,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDateTime;
-import java.util.UUID;
 import org.hibernate.annotations.GenerationTime;
 import org.hibernate.annotations.GeneratorType;